+++ src/Cwaff-Passives/LibraryCardtridge.cs
@@ -4,7 +4,7 @@ public class LibraryCardtridge : PassiveItem
 {
     public static string ItemName         = "Library Cardtridge";
     public static string ShortDescription = "Knowledge is Firepower";
-    public static string LongDescription  = "Informational items are free at shops. Bookllets are charmed upon entering a room. Piles of books explode when destroyed.";
+    public static string LongDescription  = "Books and paper-based items are free at shops. Bookllets are charmed upon entering a room. Piles of books explode when destroyed.";
     public static string Lore             = "It's pretty safe to assume that most who enter the Gungeon don't come there with the primary goal of reading books, but if you're one of the 4 who do, a library cardtridge is a must-have. Not only does it make reading that much more affordable, but when some unruly Bullet Kin inevitably swing by to destroy your preferred chair and knock over your favorite mug filled with ginger peach green tea, you'll be armed with the knowledge to transform the table you're sitting at into a true bastion of defense.";
 
     private static HashSet<int>         _BookItemIDs    = null;